Brand new 4.0.0 sgfxi, ready to roll.
Redid styles to use if [ condition ];then, otherwise staying with vertically aligned begin/end flow controls.

Added massively improved testing options to allow for other devs and branches to be tested and developed.

this is the first step in what I hope will be a major opening of this script.