new amd driver, 13.1, supports xorg 1.13, but docs say only supports kernel 3.5, can that be right?
amd once again proved that inability to do simple things like creating consistent file names indicates deeper problems
in the culture, so they once again changed the download file name. Weird, really. Almost surreal, but that's what
we are used to so it would be strange if they suddenly stopped.