attempt two for dual cards, adding in explicit busid check to handle fake dual cards.
bus id must end in .0 to be a real card id.

This is kind of a hack, so I may need to do the real lspci -n check of device type 300 and then
use that data to extract the real video devices but I need more data to do that.